List all options defined in package “modeltuner”

modeltuner_options() displays the current values of all options set in the package. These options define default values for certain function parameters.

Usage

modeltuner_options()

Value

Returns a list of option settings.

Details

cv_nfold (value at package startup: =10): Default value of argument nfold in cv(), tune() and stepwise functions.

cv_verbose (=interactive() at startup): Default value of argument verbose in cv(),

cv_show_se (=FALSE at startup): Default value of argument se in print.performance() and print.evaluation_log() and default of errorbars in plot.evaluation_log(),

cv_iter (=crit_min() at startup): Default value of argument iter in cv() applied to models of class “fm_xgb” or “fm_glmnet”,

expand_max_model (=30 at startup): Default value of argument max_n_model in multimodel.model() and tune.model(),

print_max_model (=3 at startup): Default value of n in print.multimodel(), print.cv() and print.evaluation_log(),

print_max_row (=30 at startup): Default value of n in print.performance() and print.extract_pref_iter(),

print_rows_evaluation_log (=6 at startup): Default value of n_row in print.evaluation_log().

Examples

modeltuner_options()
#> $cv_nfold
#> [1] 10
#> 
#> $cv_verbose
#> [1] FALSE
#> 
#> $cv_show_se
#> [1] FALSE
#> 
#> $cv_iter
#> Preference criterion for an iteratively fitted model:
#>   criterion:     crit_min()
#>   label suffix:  “min”
#>   Selects the iteration with minimal test error.
#> 
#> $expand_max_model
#> [1] 30
#> 
#> $print_max_model
#> [1] 3
#> 
#> $print_max_row
#> [1] 30
#> 
#> $print_rows_evaluation_log
#> [1] 6
#>