Subset an object of class “multimodel”, “cv”, “performance” or “evaluation_log”

subset() methods for classes “multimodel”, “cv” and “performance”.

Usage

# S3 method for multimodel
subset(x, subset = TRUE, ...)

# S3 method for cv
subset(x, subset = TRUE, ...)

# S3 method for performance
subset(x, subset = TRUE, ...)

# S3 method for evaluation_log
subset(x, subset = TRUE, ...)

Arguments

x

multimodel or cv or other object.

subset

Selection of models: An integer vector or a logical vector of appropriate length, or a character vector of model labels.

...

Not used.

Value

Object of same class as x.

See also

Examples

mm <- c(speed_lm = model(lm(dist ~ speed, cars)),
        speed_loess = model(loess(dist ~ speed, cars,
                                  control = loess.control(surface = "direct"))),
        speed_rpart = model(rpart::rpart(dist ~ speed, cars)))
# subset.multimodel:
subset(mm, 1:2)
#> --- A “multimodel” object containing 2 models ---
#> 
#> ‘speed_lm’:
#>   model class:  lm
#>   formula:      dist ~ speed
#>   data:         data.frame [50 x 2], 
#>                 input as: ‘data = cars’
#>   call:         lm(formula = dist ~ speed, data = data)
#> 
#> ‘speed_loess’:
#>   model class:  loess
#>   formula:      dist ~ speed
#>   data:         data.frame [50 x 2], 
#>                 input as: ‘data = cars’
#>   call:         loess(formula = dist ~ speed, data = data, control = loess.control(surface = "direct"))
# subset.cv:
cv_mm <- cv(mm)
subset(cv_mm, 3:2)
#> --- A “cv” object containing 2 validated models ---
#> 
#> Validation procedure: Complete k-fold Cross-Validation
#>   Number of obs in data:  50
#>   Number of test sets:    10
#>   Size of test sets:       5
#>   Size of training sets:  45
#> 
#> Models:
#> 
#> ‘speed_rpart’:
#>   model class:  rpart
#>   formula:      dist ~ speed
#>   metric:       rmse
#> 
#> ‘speed_loess’:
#>   model class:  loess
#>   formula:      dist ~ speed
#>   metric:       rmse
# subset.performance:
cv_perf <- cv_performance(cv_mm)
subset(cv_perf, c(1, 3))
#> --- Performance table ---
#> Metric: rmse
#>             train_rmse test_rmse time_cv
#> speed_lm        15.022    14.184   0.013
#> speed_rpart     16.543    16.679   0.029